ticket-system/py-printer/requirements.txt
altgraph==0.17.4
appdirs==1.4.4
argcomplete==3.6.2
importlib_resources==6.5.2
packaging==25.0
pillow==11.2.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
python-barcode==0.15.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.2
six==1.17.0
